Maybe fixed, maybe not - Only one way to find out! Turned GameSummary into a proper component that should only be rendered when Game Over is clicked; necessary data should be passed down as props from GameProgress; cleaned up some inconsistent data/outdated data forms in GameSummary; commented out login in GameSummary, as player should not magically be logged out but clickign GameOver.
