+++ client/src/pages/GameProgress.jsx
@@ -11,12 +11,17 @@ import { useAuthContext } from '../contexts/authContext.jsx'; //adjust if needed
 import { handleApiError } from '../utils/errorHandler.js';
 import { useSeason } from '../contexts/seasonContext.jsx'; // Import the season context
 import GameHeader from '../components/GameHeader.jsx'; // Import the GameHeader component
+import GameSummary from '../components/GameSummary.jsx'; // Import the GameSummary component
+
+
 
 export default function GameProgress() {
   const { game_id, week } = useParams(); // Get the game title from the URL parameters
   const { user } = useAuthContext(); // get the logged-in user
   const { currentSeason, setCurrentSeason, seasonThemes = {} } = useSeason(); // Access season context //FIXME: removed currentSeason = 'Spring',
   const theme = seasonThemes[currentSeason] || { bodyBg: 'bg-white', bodyText: 'text-black' }; // Get the theme based on the current season
+
+  const [showGameSummary, setShowGameSummary] = useState(false);
 
   const [game, setGame] = useState(null);
   const [stats, setStats] = useState({ abundance: '', scarcity: '', contempt: 0 });
@@ -59,7 +64,7 @@ export default function GameProgress() {
   useEffect(() => {
     console.log('Updated ongoingProjects:', ongoingProjects);
   }, [ongoingProjects]);
-  
+
   useEffect(() => {
     console.log('Updated completedProjects:', completedProjects);
   }, [completedProjects]);
@@ -76,16 +81,16 @@ export default function GameProgress() {
       const response = await projectAPI.getProjectsByGame(game_id);
       // console.log('Response from backend:', response);
       const allProjects = response.data;
-      
+
       // console.log('All projects:', allProjects);
       // Sort projects into ongoing and completed
       const ongoing = allProjects.filter(
         (proj) => proj.project_weeks > 0 || proj.pp_weeks > 0
       );
-      
+
       const completed = allProjects.filter(
         (proj) =>
-          (proj.project_weeks === 0 || proj.pp_weeks === 0) && 
+          (proj.project_weeks === 0 || proj.pp_weeks === 0) &&
           (proj.project_resolve || proj.pp_resolve)
       );
 
@@ -150,11 +155,11 @@ export default function GameProgress() {
     fetchInitialData();
   }, [game_id, currentSeason, currentWeek]);
 
-useEffect(() => {
-  if (game_id) {
-    fetchProjects();
-  }
-}, [game_id, currentWeek]); // Call fetchProjects whenever currentWeek changes
+  useEffect(() => {
+    if (game_id) {
+      fetchProjects();
+    }
+  }, [game_id, currentWeek]); // Call fetchProjects whenever currentWeek changes
 
   const fetchGameData = async () => {
     try {
@@ -230,9 +235,9 @@ useEffect(() => {
         const data = { end: formData.end }; // Assuming `formData.end` contains the epilogue
         await updateGame(game_id, data);
         console.log('Game Over data saved:', data);
-
-        // Navigate to the Game Summary page
-        navigate(`/${game_id}/summary`);
+        
+        // Show the game summary modal
+        setShowGameSummary(true);
         return; // Exit early since the game is over
       }
 
@@ -311,10 +316,10 @@ useEffect(() => {
             completedProjects={completedProjects}
             setOngoingProjects={setOngoingProjects}
             setCompletedProjects={setCompletedProjects}
-            fetchProjects={fetchProjects} 
-           />
+            fetchProjects={fetchProjects}
+          />
         </div>
-        
+
         <div className={`w-3/4`}>
           <h2 className="text-2xl font-bold mb-6 text-center">Week {currentWeek}, {currentSeason}</h2>
           {prompt && prompt._id && (
@@ -322,12 +327,12 @@ useEffect(() => {
               <div className="text-center mb-8">
                 <h3 className="text-xl font-semibold">{prompt.prompt_title}</h3>
                 <p
-                className="mb-8 leading-relaxed"
-                dangerouslySetInnerHTML={{ __html: prompt.prompt }}
+                  className="mb-8 leading-relaxed"
+                  dangerouslySetInnerHTML={{ __html: prompt.prompt }}
                 ></p>
-            </div>
+              </div>
 
-              <ActionModal 
+              <ActionModal
                 prompt={prompt}
                 game_id={game_id}
                 stats={stats}
@@ -341,24 +346,28 @@ useEffect(() => {
                 isProject={prompt?.isProject || false}
                 fetchProjects={fetchProjects}
                 setOngoingProjects={setOngoingProjects}
-                />
-
-
-            <div className="text-center">
-              <button
-                className={`btn mt-6 shadow-md border-none ${theme.nextWeekBtnBg} ${theme.nextWeekBtnText} ${theme.nextWeekBtnBgHover}`}
-                onClick={handleNextWeek}
-              >
-                {prompt._id.toString() === GAME_OVER_PROMPT_ID ? 'Game Over' : 'Next Week'}
-              </button>
-              <GameSummary
-                game={game}
-                stats={stats}
-                projects={projects}
-                currentWeek={currentWeek}
-                loading={loading}
               />
-            </div>
+
+
+              <div className="text-center">
+                <button
+                  className={`btn mt-6 shadow-md border-none ${theme.nextWeekBtnBg} ${theme.nextWeekBtnText} ${theme.nextWeekBtnBgHover}`}
+                  onClick={handleNextWeek}
+                >
+                  {prompt._id.toString() === GAME_OVER_PROMPT_ID ? 'Game Over' : 'Next Week'}
+                </button>
+                {showGameSummary && (
+                  <GameSummary
+                    className="bg-white p-6 rounded-lg shadow-lg max-w-md w-full"
+                    onClose={() => setShowGameSummary(false)}
+                    game={game}
+                    stats={stats}
+                    projects={projects}
+                    currentWeek={currentWeek}
+                    loading={loading}
+                  />
+                )}
+              </div>
             </div>
           )}
         </div>
